use consistent "outer/inner" naming in all position and dimension properties.
In this viewport naming, the viewport is synonymous with the container element and position/dimensions are calculated from it.

"outer" values are calculated from the perspective of the document's absolute positions (ie. after scaling). This matches how they appear onscreen in actuality. These are used when collecting positions of the mouse and outer elements.
"inner" values are represented relative to the container's position ignoring the scaling.  This is necessary when the result of the calculation is displayed inside the container and subject to its scaling (indirectly).